Update bangenum.sh
Assigned variables to certain commands for later calls. Added output check to tell what specific architecture is as I usually end up referring to my notes or Googling the output of 'uname -i'. I also added a check to tell you if the 'OutOfBox' account is present, as I recently stumbled across the account in VHL.